Add comprehensive tests for pipeline, query, and storage operations
- Implement tests for the pipeline runner, including cleanup and run cycle functionalities.
- Add tests for vector search operations, ensuring correct query structure and parameter handling.
- Create tests for article storage operations, covering upsert, retrieval, and update functionalities.
- Introduce tests for cluster storage operations, validating cluster creation, assignment, and retrieval.
- Establish tests for database connection pool management, ensuring proper initialization and closure.
- Implement tests for pipeline state tracking operations, verifying state retrieval and updates.
