+++ docs/design/plans/2026-04-20-api-pipeline-split.md
