+++ docs/design/plans/2026-04-23-api-key-service.md
