+++ docs/design/specs/2026-04-20-api-pipeline-split-design.md
