+++ docs/design/specs/2026-04-21-phase2-design.md
