+++ src/gdelt_event_pipeline/api/static/archive/globe.html
