--- src/gdelt_event_pipeline/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Shared utilities."""
