+++ tests/api/test_app.py
@@ -1,42 +1,32 @@
 """Tests for app.py behaviour specific to the Vercel deployment split."""
 
-from unittest.mock import patch
+from unittest.mock import MagicMock, patch
 
-import pytest
 from fastapi.testclient import TestClient
 
 import gdelt_event_pipeline.api.app as app_module
-
-
-@pytest.fixture
-def client_no_db():
-    """TestClient with DB lifecycle mocked out so no real connection is needed."""
-    with (
-        patch("gdelt_event_pipeline.api.app.init_pool"),
-        patch("gdelt_event_pipeline.api.app.close_pool"),
-        patch("gdelt_event_pipeline.api.app._ensure_schema"),
-    ):
-        with TestClient(app_module.app) as client:
-            yield client
+import gdelt_event_pipeline.api.middleware as middleware_module
+import gdelt_event_pipeline.api.routers.search as search_module
+from tests.conftest import make_mock_pool
 
 
 class TestSearchGuard:
     def test_search_returns_501_when_search_unavailable(self, client_no_db, monkeypatch):
         """When _SEARCH_AVAILABLE is False the endpoint must return HTTP 501."""
-        monkeypatch.setattr(app_module, "_SEARCH_AVAILABLE", False)
+        monkeypatch.setattr(search_module, "_SEARCH_AVAILABLE", False)
         response = client_no_db.get("/api/search?q=test")
         assert response.status_code == 501
         assert "not available" in response.json()["detail"].lower()
 
     def test_search_available_flag_true_in_full_environment(self):
         """fastembed is in the main dependencies, so the module-level flag must be True."""
-        assert app_module._SEARCH_AVAILABLE is True
+        assert search_module._SEARCH_AVAILABLE is True
 
     def test_search_returns_501_when_search_available_flag_is_false(
         self, client_no_db, monkeypatch
     ):
         """When _SEARCH_AVAILABLE is False, /api/search must return 501 regardless of backend."""
-        monkeypatch.setattr(app_module, "_SEARCH_AVAILABLE", False)
+        monkeypatch.setattr(search_module, "_SEARCH_AVAILABLE", False)
         response = client_no_db.get("/api/search?q=test")
         assert response.status_code == 501
 
@@ -48,7 +38,7 @@ class TestServerlessPoolSizing:
         with (
             patch("gdelt_event_pipeline.api.app.init_pool") as mock_init,
             patch("gdelt_event_pipeline.api.app.close_pool"),
-            patch("gdelt_event_pipeline.api.app._ensure_schema"),
+            patch("gdelt_event_pipeline.api.app.ensure_schema"),
         ):
             with TestClient(app_module.app):
                 pass
@@ -62,7 +52,7 @@ class TestServerlessPoolSizing:
         with (
             patch("gdelt_event_pipeline.api.app.init_pool") as mock_init,
             patch("gdelt_event_pipeline.api.app.close_pool"),
-            patch("gdelt_event_pipeline.api.app._ensure_schema"),
+            patch("gdelt_event_pipeline.api.app.ensure_schema"),
         ):
             with TestClient(app_module.app):
                 pass
@@ -74,16 +64,13 @@ class TestServerlessPoolSizing:
 class TestRateLimiting:
     def test_redis_rate_limit_returns_429_when_over_limit(self, client_no_db, monkeypatch):
         """When Redis pipeline reports count >= RATE_LIMIT_MAX, middleware returns 429."""
-        from unittest.mock import MagicMock
-
         mock_pipe = MagicMock()
-        # Pipeline results: [zremrangebyscore_result, zcard_count, zadd_result, expire_result]
-        mock_pipe.execute.return_value = [0, 30, 1, 1]  # count=30 == RATE_LIMIT_MAX
+        mock_pipe.execute.return_value = [0, 30, 1, 1]
 
         mock_redis = MagicMock()
         mock_redis.pipeline.return_value = mock_pipe
 
-        monkeypatch.setattr(app_module, "_redis", mock_redis)
+        monkeypatch.setattr(middleware_module, "_redis", mock_redis)
 
         response = client_no_db.get("/api/clusters")
         assert response.status_code == 429
@@ -91,62 +78,45 @@ class TestRateLimiting:
 
     def test_redis_rate_limit_passes_when_under_limit(self, client_no_db, monkeypatch):
         """When Redis pipeline reports count < RATE_LIMIT_MAX, middleware passes through."""
-        from unittest.mock import MagicMock
-
         mock_pipe = MagicMock()
-        mock_pipe.execute.return_value = [0, 5, 1, 1]  # count=5, well under limit
+        mock_pipe.execute.return_value = [0, 5, 1, 1]
 
         mock_redis = MagicMock()
         mock_redis.pipeline.return_value = mock_pipe
 
-        monkeypatch.setattr(app_module, "_redis", mock_redis)
-        # /api/search with _SEARCH_AVAILABLE=False returns 501 — proves middleware passed through
-        monkeypatch.setattr(app_module, "_SEARCH_AVAILABLE", False)
+        monkeypatch.setattr(middleware_module, "_redis", mock_redis)
+        monkeypatch.setattr(search_module, "_SEARCH_AVAILABLE", False)
 
         response = client_no_db.get("/api/search?q=test")
-        assert response.status_code == 501  # middleware passed through; endpoint returned 501
+        assert response.status_code == 501
 
     def test_falls_back_to_in_memory_when_redis_not_configured(self, client_no_db, monkeypatch):
         """When _redis is None (no Upstash credentials), middleware uses in-memory store."""
-        monkeypatch.setattr(app_module, "_redis", None)
-        monkeypatch.setattr(app_module, "_SEARCH_AVAILABLE", False)
+        monkeypatch.setattr(middleware_module, "_redis", None)
+        monkeypatch.setattr(search_module, "_SEARCH_AVAILABLE", False)
 
         response = client_no_db.get("/api/search?q=test")
-        # In-memory fallback allows the request through (count is 0)
-        assert response.status_code == 501  # passed through; not 429
+        assert response.status_code == 501
 
 
 class TestApiKeyAuth:
     def test_no_key_header_passes_with_anonymous_limit(self, client_no_db, monkeypatch):
         """Requests with no X-API-Key go through with the anonymous rate limit."""
-        monkeypatch.setattr(app_module, "_redis", None)
-        monkeypatch.setattr(app_module, "_SEARCH_AVAILABLE", False)
+        monkeypatch.setattr(middleware_module, "_redis", None)
+        monkeypatch.setattr(search_module, "_SEARCH_AVAILABLE", False)
 
         response = client_no_db.get("/api/search?q=test")
-        assert response.status_code == 501  # middleware passed; endpoint returned 501
+        assert response.status_code == 501
 
     def test_invalid_key_returns_401(self, client_no_db, monkeypatch):
         """When X-API-Key does not match any active key in DB, return 401."""
-        from unittest.mock import MagicMock
+        monkeypatch.setattr(middleware_module, "_redis", None)
 
-        monkeypatch.setattr(app_module, "_redis", None)
-
-        mock_cur = MagicMock()
-        mock_cur.__enter__ = lambda s: s
-        mock_cur.__exit__ = MagicMock(return_value=False)
-        mock_cur.fetchone.return_value = None  # key not found
-
-        mock_conn = MagicMock()
-        mock_conn.__enter__ = lambda s: s
-        mock_conn.__exit__ = MagicMock(return_value=False)
-        mock_conn.cursor.return_value = mock_cur
-
-        mock_pool = MagicMock()
-        mock_pool.connection.return_value = mock_conn
+        pool = make_mock_pool(fetchone_return=None)
 
         import gdelt_event_pipeline.storage.database as db_module
 
-        with patch.object(db_module, "get_pool", return_value=mock_pool):
+        with patch.object(db_module, "get_pool", return_value=pool):
             response = client_no_db.get(
                 "/api/search?q=test", headers={"X-API-Key": "gdp_invalidkey"}
             )
@@ -156,13 +126,11 @@ class TestApiKeyAuth:
 
     def test_auth_endpoints_skip_middleware(self, client_no_db):
         """/api/auth/* paths bypass the key check and rate limiter."""
-        # /api/auth/keys requires a valid Clerk JWT — without one it returns 422 (missing Header)
-        # but NOT 401 from the middleware, proving middleware was skipped
         response = client_no_db.get("/api/auth/keys")
-        assert response.status_code == 422  # FastAPI schema validation, not 401 from middleware
+        assert response.status_code == 422
 
     def test_static_pages_not_protected(self, client_no_db):
-        """Static routes (/, /globe, etc.) must be accessible without any API key."""
+        """Static routes (/, etc.) must be accessible without any API key."""
         response = client_no_db.get("/")
         assert response.status_code == 200
